refactor: optional 등 dto 내에서 private 메서드로 분리하도록 수정 (#20)
